Modify rule grammar
This patch creates a new rule grammar that adds a new permission request
field, perm, to the rules. The perm field may be omitted for backwards
compatibility which will assume the request is for opening a file. This
patch also exposes the trust database though subject and object fields.
These are subj_trust and obj_trust respectively. These may be either a
0 or 1 to inidicate that the item being in the trust database is or is not
required.