Make the watched filesystems configurable
Currently, the filesystems that are watched are hardcoded to be xfs,tmpfs,
ext4,ext3,ext2. This is not configurable. So, if someone needs to watch
something based on dosfs, they have no way to do it.

This change introduces a new fapolicyd.conf option, watch_fs, so that the
end user may specify precisely which filesystems to watch. This list is
used for dynamic mountpoint inclusion when a hotplug mount event occurs.

The design is to use a AVL tree since loading the filesystems occurs once
and then it may need to be searched many times. An AVL tree should give
superior performance compared to a simple linked list.