ELF file detection
Turns out libmagic is very unpredictable in determining elf file types
between OS releases and based on pie or other attributes. Since we
already have an elf parser that collect artifacts within the file, just
use its output to determine type.

On non-elf files, this should end fast because gather_elf looks
for ELFMAGIC in first few bytes and exits if not found. In that case
we just continue like we used to.