Redesign rule syntax to expose subject and object trust
In a previous patch, subject and object trust was cached for re-use.
This lead to an observartion that this could be directly exposed
to the rules. This could be used to increase performance since previous
rules have a lot of path checking.

The rule syntax was also redesigned to have a colon ':' between the
subject and object side. This allows better perdiction of when
'all' is going to be considered either subject or object. Now it is
explicit. See the man page and the following issue for more discussion:

https://github.com/linux-application-whitelisting/fapolicyd/issues/47