Update rules
The comments were updated to better reflect what the policy is. Lua
scripting was previously missed and it should be blocked by default.
So, comment was added about when perl might be needed. And a python
rule had a glob that was not needed. After some thought, 4 python
rules were combined into 2. The first two were the same as perm=any.
The next two allow open of trusted python scripts. The only difference
is they differ by exe. In the end, it doesn't matter which application
opens them as they are trusted.