Fix analysis
The code that is removed by this patch is leftover code from previous
pattern matching rules. It is unnecessary and causes some cases to
not be changed from STATE_FULL to its real status. Then when it leaves
the function, proc_info gets deleted so that next time it segfaults
checking path1 because state is once again STATE_FULL.