Add ability to define the format of the syslog event
With this commit, the admin now has the ability to define the format
of the event by altering the syslog_format configuration item. The
format will also be used for output in debug mode. This lets one fine tune
how it looks before using it in the logs. The audit event is unchanged
because the kernel completely handles that.