Correct denied accounting
The code currently uses an inverted AUDIT mask to pick out the decision.
This was fine when there was only audit. Now there is also syslog. That
means that any that have syslog will be consider allowed even if the
decision is to deny. Correct this by checking the actual deny bit to
see if its a denial or not.