Add Group ID support for rules
There has been a desire to add support for writing rules that use
the Group ID to make a decision. This patch makes it possible to
write rules like:

allow perm=any gid=wheel : all

This ability should be used sparingly as this could lead to policy that
is more like an access control system than a reputation based control
system.