Supplemental Group Support
The previous patch adding Group ID support did not use supplemental
group information. This patch adds that support by picking up extra
information in /proc<pid>/status and turning that itself into a "set".
This means that to determine a match, we need to find the intersection
of 2 AVL trees, one being the rule and the other being the process
attributes.

This patch also adds a test program to verify that the low level parts
work correctly. The only known problem left unsolved by this patch is
printing the group information to syslog when so configured. This will
be addressed in the future as it requires modification to the basic
AVL data structures.