Add support for rules.d directory (#162)
Currently fapolicyd has its rules in one monolithic rules file. This
makes it hard to add units of policy based on software installed on the
machine. This patch adds a rules.d directory and a utility that compiles
the master rules file out of component files in rules.d.

It maintains backwards compatibility with the old rules file. If the new
compiled.rules file is not found, it opens the old one if it exists.