Fix for CVE-2022-1117
The build script makes an assumption how glibc names it's runtime linker.
This worked fine until glibc-2.34. Starting with that, the naming
convention changed from /usr/lib64/ld-2.*.so to /lib64/ld-linux-x86-64.so.2
The upshot of this is the build script misdetects the run time linker
and this causes the ld_so pattern matcher to not work correctly.
Additionally, non-x86_64 platforms may not have followed the ld-2.*.so
pattern and were misdeteced prior to glibc-2.34.

Would like to thank NCCGroup for reporting this problem.