Remove update thread locking
The main thread does not interact with the trust database. This means
it should not be locking/unlocking access to the trust database. The
main thread only needs to lock the inbound queue so it stays in sync
with the decision thread. The decision thread is the one that needs to
synchronize with the update thread.