Reduce locking in the decision thread
While the decision thread has the lock of the inbound queue, check
to see if there are more events and handle up to MAX_EVENTS per
lock. If there is contention around the inbound queue, this should
help. Currently, MAX_EVENTS is 4. The value was randomly chosen. But
since it is making an array on the stack, let's keep it smallish.