Speed up sha256 hashing
The current algorithm reads the file 4K bytes at a time. It loops until
EOF and then asks libgcrypt for the final hash. This seems to be where
fapolicyd spends about 27% of it's time when answering if access is
allowed or not.

Various sized buffers were tested to see if a speed could be made with
the following speed improvements:

4K:  0%
8K:  4%
16K: 6%
32K: 8%
64K: 8%

Buffer alignment was checked and found 64k aligned was 10% faster. Next
it was checked to see if the mmap and gcry_md_hash_buffer yielded better
results. This was found to be 13% faster. Next, MAP_POPULATE was added
as a flag. It is supposed to start filling the memory so that when
gcry_md_hash_buffer is called, there is less time waiting on faults.
This turned out to be 14% faster.

Openssl's SHA256 was also checked to see if it was faster. No difference
could be measured. I think this patch makes sha256 hashing as fast as
we can get it without assembler changes in libgcrypt.