Adjust cli app to new sha256 function
The cli is misdetecting many files sha256 as a mismatch. Turns out that
a lot have a length of 0. We cannot has files of no size, so avoid that.
Right be fore the hash, we do a size check. So, we'll depend on that
to detect something has changed. Also, if the hash function has an
error, warn about a miscompare as a precaution.