Speedup logging to stderr
When using a tool like Helgrind to check concurrency issues, all writes
to stderr involve locking and unlocking the tty. Convert all the fputc
loops to just use fputs so there is only one lock/unlock instead of many.
This improves the performance overall because it can't make decisions if
is locking and unlocking the tty.