Fix max_depth reporting
When sending SIGUSR1 to fapolicyd, a lot of the time the statistics are
saying the Inter-thread max queue depth is 0. It has to be at least 1 to
be valid results. This comes from having 2 max_depth that are only
synchronized on exit. That means reports while it's running are incorrect.
This is fixed by eliminating one copy so that they don't need synchronizing.