Introduction of rpmdb transaction locking
We need to use rpm transaction lock so we can be sure we are reading
valid(not corrupted) data.

If we cannot acquire lock, we have to stay with an old version of trustdb.
We expect that files installed in current transaction are present
in the old trustdb already, they were recieved from fapoliycd rpm plugin.
So it is safe to wait till the end of the transaction and then trustdb
will be reloaded.

Signed-off-by: Radovan Sroka <rsroka@redhat.com>