Fix null argv when spawning fapolicyd-rpm-loader
On startup the kernel prints a warning that:

process 'fapolicyd' launched '/usr/sbin/fapolicyd-rpm-loader' with NULL argv: empty string added

Conventionally the first argument to argv should be the name of the
program, and the kernel prints a warning whenever this is not the case.

https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=dcd46d897adb70d63e025f175a00a89797d31a43

fapolicyd-rpm-loader doesn't use any of its arguments, but add an empty
string to argv anyway to silence the kernel warning.

Signed-off-by: Jacob Garber <jacob@jwgarber.ca>