Drop device keyword for subjects
I don't know if it ever made sense to write rules based on a subject's
original device. If it did, we should have caught it when execution
began and it was still an object. The code has always had a FIXME because
of lingering questions about the value of this. We are now dropping
the idea and have one less FIXME.