Revert commit 643fdce91be22504b53e3f6cb2c5e1842d391095
The patch fixed the leaks by updating the filter code. The effect is that
the leaks went away, but caused problems in the filter. Suspect the leaks
should be fixed in the filter_test program. They appeared to originate
in the filter code:

Indirect leak of 1333 byte(s) in 173 object(s) allocated from:
    #0 0x7f01082e0ca0 in strdup (/lib64/libasan.so.8+0xe0ca0) (BuildId: 10b8ccd49f75c21babf1d7abe51bb63589d8471f)
    #1 0x00000040427c in filter_load_file library/filter.c:556

Will investigate the filter_test program.