Fix uninitialzed subject array element
Removing the EXE_DEVICE uncovered a long standing bug. There has always
been an off by one in initializing the subject array. SUBJ_START occupies
a slot. It probably should be a define pointing to AUID. But it occupies
a slot that is unused. There are loops that use < SUBJ_END which points
to EXE_TYPE. So, everything less that EXE_TYPE is initialized. But EXE_TYPE
is used and produces a segfault. Previously, it would have been EXE_DEVICE
which was unused.

Turn the calculation in to a macro and use it in the for loops.