Fix uninitialzed object array element
There has always been an off by one in initializing the object array.
OBJ_START occupies a slot. It probably should be a define pointing to PATH.
But it occupies a slot that is unused. There are loops that use < OBJ_END
which points to FMODE. So, everything less that FMODE is initialized.
But FMODE is used and does not produces a segfault. Therefore it is
undetected. If someone ever wrote a rule to use it, the code would segfault.