Add --test-filter to fapolicy-cli
There currently is no way to check the filter. This new option
takes a path to a file and replies with allow or deny to let you
know if the filter is working right. This can be handy when adding
more rules to fine tune what's in the trust database.

This also updates the bash_completions and man page.