cli: add --check-ignore_mounts option
add --check-ignore_mounts option that can optionally take a target mount point on the comand line to override the ignore_mounts configuration. Otherwise it parses ignore_mounts, walks each directory to look for files that could be executed, and exits non-zero when %languages MIME matches are found. It is important to highlight this concern because when the mount is ignored and has executable material, it cannot be controlled by fapolicyd.