Introduce read_proc_status()
The .proc/<pid>/status file has information that we need. But it currently
is used piece meal. If we need the gid, we go open it. If we need the uid,
we go open it. If we need comm, we open a different file. So, the idea was
that during rule parsing, build a bitmap for each field that should be
collected from the status page. Then get them all at once so that we don't
keep opening and parsing the page.