Allow fapolicyd to have other paths
In fapolicyd-cli, when doing the check-status command, it checks to
see if the application pointed to be fapolicyd.pid really is fapolicyd
by checking the complete path against the hardcoded path. This prevents
getting a status report if it is started from somewhere else.

Adjust the check to only check the basename of the application. That will
allow execution from other places.