Add support for kworker threads
Kworker threads run external programs by using call_usermodehelper()
in the kernel. They are unique in that they have no /proc/<pid>/exe
file. They otherwise wind up in a STATE_NORMAL posture. Support them
by simply checking if the exe entry is missing and diverting to COMM
if so. Since they are never an object, gather_elf is not called and
everything just works (now that we have the proper process name).