Generalize hashing infrastructure and rename cache slot
Update file hash functions and attributes to support multiple hash algorithms.
The sha256hash object used in rules will now be file_hash. The algorithm
used will be detected from the hash length. It will continue to support
sha256hash in rules for now. This generalization is in preparation to support
sha512 hashes in rpms.