Get rid of unreliable detections
SVG is problematic because the <svg can be after a long copyright or license
notice which is variable in size. So, that means we have to get rid of the
XML detection in general since svg is a specializaion of XML.

Also drop the python bytecode detection. It was too general and called too
many things python bytecode when it was not.